# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: RISE 성과관리 시스템 API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /projects:
    get:
      operationId: listProjects
      tags: [projects]
      summary: 프로젝트 목록
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Project"
    post:
      operationId: createProject
      tags: [projects]
      summary: 프로젝트 생성
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProjectBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
  /projects/{id}:
    get:
      operationId: getProject
      tags: [projects]
      summary: 프로젝트 조회
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "404":
          description: Not Found
    patch:
      operationId: updateProject
      tags: [projects]
      summary: 프로젝트 수정
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProjectBody"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
    delete:
      operationId: deleteProject
      tags: [projects]
      summary: 프로젝트 삭제
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: No Content

  # ── Tasks ─────────────────────────────────────────────────
...
  /indicators:
    get:
      operationId: listIndicators
      tags: [indicators]
      summary: 지표 목록
      parameters:
        - name: taskId
          in: query
          schema:
            type: integer
        - name: parentId
          in: query
          schema:
            type: ["integer", "null"]
        - name: indicatorType
          in: query
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Indicator"
    post:
      operationId: createIndicator
      tags: [indicators]
      summary: 지표 생성
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateIndicatorBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Indicator"
  /indicators/{id}:
    get:
      operationId: getIndicator
      tags: [indicators]
      summary: 지표 조회
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Indicator"
        "404":
          description: Not Found
    patch:
      operationId: updateIndicator
      tags: [indicators]
      summary: 지표 수정
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateIndicatorBody"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Indicator"
    delete:
      operationId: deleteIndicator
      tags: [indicators]
      summary: 지표 삭제
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: No Content

  # ── Indicator Targets ─────────────────────────────────────
...
  /results:
    get:
      operationId: listResults
      tags: [results]
      summary: 실적 목록
      parameters:
        - name: indicatorId
          in: query
          schema:
            type: integer
        - name: year
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/IndicatorResult"
    post:
      operationId: createResult
      tags: [results]
      summary: 실적 입력
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateResultBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IndicatorResult"
  /results/{id}:
    get:
      operationId: getResult
      tags: [results]
      summary: 실적 조회
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IndicatorResult"
        "404":
          description: Not Found
    patch:
      operationId: updateResult
      tags: [results]
      summary: 실적 수정
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateResultBody"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IndicatorResult"
    delete:
      operationId: deleteResult
      tags: [results]
      summary: 실적 삭제
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: No Content
  /results/{id}/submit:
    post:
      operationId: submitResult
      tags: [results]
      summary: 실적 제출
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IndicatorResult"

  # ── Evidence Files ─────────────────────────────────────────
...
  /reviews:
    get:
      operationId: listReviews
      tags: [reviews]
      summary: 검토 목록
      parameters:
        - name: resultId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Review"
    post:
      operationId: createReview
      tags: [reviews]
      summary: 검토 등록
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateReviewBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Review"

  # ── Feedback Actions ───────────────────────────────────────
...
components:
  schemas:
...
    Project:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        startYear:
          type: integer
        endYear:
          type: integer
        status:
          type: string
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - name
        - startYear
        - endYear
        - status
        - createdAt
        - updatedAt
    CreateProjectBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: ["string", "null"]
        startYear:
          type: integer
        endYear:
          type: integer
        status:
          type: string
      required:
        - name
        - startYear
        - endYear
    UpdateProjectBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: ["string", "null"]
        startYear:
          type: integer
        endYear:
          type: integer
        status:
          type: string

    # ── Tasks ──────────────────────────────────────────────
...
    Indicator:
      type: object
      properties:
        id:
          type: integer
        taskId:
          type: integer
        parentId:
          type: ["integer", "null"]
        indicatorType:
          type: string
          description: "parent | child"
        name:
          type: string
        unit:
          type: ["string", "null"]
        formula:
          type: ["string", "null"]
        weight:
          type: ["number", "null"]
        description:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - taskId
        - indicatorType
        - name
        - createdAt
        - updatedAt
    CreateIndicatorBody:
      type: object
      properties:
        taskId:
          type: integer
        parentId:
          type: ["integer", "null"]
        indicatorType:
          type: string
        name:
          type: string
        unit:
          type: ["string", "null"]
        formula:
          type: ["string", "null"]
        weight:
          type: ["number", "null"]
        description:
          type: ["string", "null"]
      required:
        - taskId
        - indicatorType
        - name
    UpdateIndicatorBody:
      type: object
      properties:
        parentId:
          type: ["integer", "null"]
        indicatorType:
          type: string
        name:
          type: string
        unit:
          type: ["string", "null"]
        formula:
          type: ["string", "null"]
        weight:
          type: ["number", "null"]
        description:
          type: ["string", "null"]

    # ── Indicator Targets ──────────────────────────────────
...
    IndicatorResult:
      type: object
      properties:
        id:
          type: integer
        indicatorId:
          type: integer
        year:
          type: integer
        actualValue:
          type: ["number", "null"]
        calculatedValue:
          type: ["number", "null"]
        progressRate:
          type: ["number", "null"]
        status:
          type: string
          description: "draft | submitted | reviewing | revision_requested | approved | rejected"
        selfEvaluation:
          type: ["string", "null"]
        submittedAt:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - indicatorId
        - year
        - status
        - createdAt
        - updatedAt
    CreateResultBody:
      type: object
      properties:
        indicatorId:
          type: integer
        year:
          type: integer
        actualValue:
          type: ["number", "null"]
        selfEvaluation:
          type: ["string", "null"]
      required:
        - indicatorId
        - year
    UpdateResultBody:
      type: object
      properties:
        actualValue:
          type: ["number", "null"]
        selfEvaluation:
          type: ["string", "null"]
        status:
          type: string

    # ── Evidence Files ─────────────────────────────────────
...
    Review:
      type: object
      properties:
        id:
          type: integer
        resultId:
          type: integer
        reviewerName:
          type: string
        reviewStatus:
          type: string
          description: "approved | revision_requested | rejected"
        comment:
          type: ["string", "null"]
        reviewedAt:
          type: string
        createdAt:
          type: string
      required:
        - id
        - resultId
        - reviewerName
        - reviewStatus
        - reviewedAt
        - createdAt
    CreateReviewBody:
      type: object
      properties:
        resultId:
          type: integer
        reviewerName:
          type: string
        reviewStatus:
          type: string
        comment:
          type: ["string", "null"]
      required:
        - resultId
        - reviewerName
        - reviewStatus

    # ── Feedback Actions ───────────────────────────────────
...